+++ noodles-vcf/src/header/format.rs
@@ -2,9 +2,10 @@
 
 mod builder;
 pub mod key;
+mod tag;
 pub mod ty;
 
-pub use self::{key::Key, ty::Type};
+pub use self::{key::Key, tag::Tag, ty::Type};
 
 use std::{error, fmt, num};
 
@@ -27,7 +28,7 @@ pub struct Format {
     ty: Type,
     description: String,
     idx: Option<usize>,
-    fields: IndexMap<String, String>,
+    fields: IndexMap<tag::Other, String>,
 }
 
 impl Format {
@@ -177,7 +178,7 @@ impl Format {
     /// let format = Format::from(Key::Genotype);
     /// assert!(format.fields().is_empty());
     /// ```
-    pub fn fields(&self) -> &IndexMap<String, String> {
+    pub fn fields(&self) -> &IndexMap<tag::Other, String> {
         &self.fields
     }
 }
@@ -274,25 +275,25 @@ fn parse_struct(
     let mut builder = Builder::default();
 
     for (key, value) in fields {
-        builder = match key.as_ref() {
-            ID => {
+        builder = match Tag::from(key) {
+            tag::ID => {
                 let id = value.parse().map_err(TryFromRecordError::InvalidId)?;
                 builder.set_id(id)
             }
-            NUMBER => {
+            tag::NUMBER => {
                 let number = value.parse().map_err(TryFromRecordError::InvalidNumber)?;
                 builder.set_number(number)
             }
-            TYPE => {
+            tag::TYPE => {
                 let ty = value.parse().map_err(TryFromRecordError::InvalidType)?;
                 builder.set_type(ty)
             }
-            DESCRIPTION => builder.set_description(value),
-            IDX => {
+            tag::DESCRIPTION => builder.set_description(value),
+            tag::IDX => {
                 let idx = value.parse().map_err(TryFromRecordError::InvalidIdx)?;
                 builder.set_idx(idx)
             }
-            _ => builder.insert(key, value),
+            Tag::Other(t) => builder.insert(t, value),
         };
     }
 
@@ -400,7 +401,7 @@ mod tests {
     }
 
     #[test]
-    fn test_try_from_record_for_format_with_extra_fields() {
+    fn test_try_from_record_for_format_with_extra_fields() -> Result<(), &'static str> {
         let record = Record::new(
             record::Key::Format,
             record::Value::Struct(vec![
@@ -421,11 +422,16 @@ mod tests {
                 ty: Type::String,
                 description: String::from("Genotype"),
                 idx: Some(1),
-                fields: [(String::from("Comment"), String::from("noodles"))]
-                    .into_iter()
-                    .collect(),
+                fields: [(
+                    Tag::other("Comment").ok_or("invalid tag")?,
+                    String::from("noodles")
+                )]
+                .into_iter()
+                .collect(),
             })
         );
+
+        Ok(())
     }
 
     #[test]